agents/visualization-agent/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
plotly==5.18.0
pandas==2.1.4
google-cloud-storage==2.14.0
pydantic==2.5.0
